# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "User Registration API",
        "description": "User Registration API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "node-express-user-registration.herokuapp.com",
    "tags": [
        {
            "name": "Users",
            "description": "API for User Registration"
        }
    ],
    "schemes": [
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/users": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create user",
                "description": "Create new user",
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "description": "User that you want to create",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "New user is created",
                        "schema": {
                            "$ref": "#/definitions/CreatedUser"
                        }
                    },
                    "400": {
                        "description": "Email is taken",
                        "schema": {
                            "$ref": "#/definitions/BusinessResponse"
                        }
                    },
                    "400": {
                        "description": "Error while checking is user exists.",
                        "schema": {
                            "$ref": "#/definitions/BusinessResponse"
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Users"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/BusinessResponse"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "summary": "User login with given credentials",
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "description": "User with email and password",
                        "schema": {
                            "$ref": "#/definitions/LoginCredentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User login succesfully",
                        "schema": {
                            "$ref": "#/definitions/BusinessResponse"
                        }
                    },
                    "400": {
                        "description": "Error while login",
                        "schema": {
                            "$ref": "#/definitions/BusinessResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid email or password",
                        "schema": {
                            "$ref": "#/definitions/BusinessResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "User": {
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "uniqueItems": true
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "CreatedUser": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "Users": {
            "type": "array",
            "$ref": "#/definitions/CreatedUser"
        },
        "LoginCredentials": {
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "BusinessResponse": {
            "properties": {
                "status": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}